runtime: fix rewindmorestack

+++ src/runtime/sys_arm64.go
@@ -24,9 +24,9 @@ func rewindmorestack(buf *gobuf) {
 	var inst uint32
 	if buf.pc&3 == 0 && buf.pc != 0 {
 		inst = *(*uint32)(unsafe.Pointer(buf.pc))
-		// section C6.6.20 Unconditional branch (immediate)
+		// section C3.2.6 Unconditional branch (immediate)
 		if inst>>26 == 0x05 {
-			buf.pc += uintptr(int64(inst<<6)>>4) + 8
+			buf.pc += uintptr(int32(inst<<6)>>4) + 8
 			return
 		}
 	}
